requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-template-maths==0.2.0
Pillow==9.2.0
psycopg==3.1.6
psycopg-binary==3.1.6
psycopg2-binary==2.9.3
sqlparse==0.4.2
typing_extensions==4.4.0
